pyproject.toml
[project]
name = "mcp-atlassian"
version = "0.1.0"
description = "The Model Context Protocol (MCP) Atlassian integration is an open-source implementation that bridges Atlassian products (Jira and Confluence) with AI language models following Anthropic's MCP specification. This project enables secure, contextual AI interactions with Atlassian tools while maintaining data privacy and security. Key features include:"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "atlassian-python-api>=3.41.16",
 "beautifulsoup4>=4.12.3",
 "httpx>=0.28.0",
 "mcp>=1.0.0",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "sooperset"
email = "soomiles.dev@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-atlassian = "mcp_atlassian:main"
